break circular dependency on ALParser. Moved static functions form ALObject.
